add stock, store *** update address-user association

+++ backend/src/main/java/com/example/bookstore/address/AddressRepo.java
@@ -4,7 +4,9 @@ import com.example.bookstore.user.User;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import javax.transaction.Transactional;
+
 @Repository
-public interface AddressRepo extends CrudRepository<Address, Integer> {
+public interface AddressRepo extends CrudRepository<Address, Long> {
     Address findByUser(User user);
 }
